+++ frend/reporting/src/app/areas/electron-nodejs/extra-packages/extra-packages.template.ts
@@ -13,10 +13,36 @@ export const extraPackagesTemplate = `<!-- <ng-template #extraPackagesTemplate> 
               >&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;
               @if (extraPackage.status == 'not-installed') {
                 <em>{{ 'AREAS.INSTALL-SETUP-UPGRADE.COMPONENTS.EXTRA-PACKAGES.INSTALLED.NOT' | translate }}</em>
+                &nbsp;&nbsp;
+                <button
+                  id="btnInstallPackage-{{extraPackage.id}}"
+                  type="button"
+                  class="btn btn-outline btn-primary btn-xs align-middle"
+                  [disabled]="extraPackage.busy"
+                  (click)="installPackage(extraPackage)"
+                >
+                  @if (extraPackage.busy) {
+                    <span class="loading loading-spinner loading-sm"></span>
+                  }
+                  {{ 'AREAS.INSTALL-SETUP-UPGRADE.COMPONENTS.EXTRA-PACKAGES.ACTIONS.INSTALL-NOW' | translate }}
+                </button>
               }
               @if (extraPackage.status != 'not-installed') {
                 <strong>{{ 'AREAS.INSTALL-SETUP-UPGRADE.COMPONENTS.EXTRA-PACKAGES.INSTALLED.ALREADY' | translate }}</strong>
                 <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 24 24" fill="#16a34a" class="inline-block w-5 h-5" style="vertical-align:middle; margin-left:4px;" aria-label="installed"><path fill-rule="evenodd" d="M2.25 12c0-5.385 4.365-9.75 9.75-9.75s9.75 4.365 9.75 9.75-4.365 9.75-9.75 9.75S2.25 17.385 2.25 12Zm13.36-1.814a.75.75 0 1 0-1.22-.872l-3.236 4.53L9.53 12.22a.75.75 0 0 0-1.06 1.06l2.25 2.25a.75.75 0 0 0 1.14-.094l3.75-5.25Z" clip-rule="evenodd"/></svg>
+                &nbsp;&nbsp;
+                <button
+                  id="btnUninstallPackage-{{extraPackage.id}}"
+                  type="button"
+                  class="btn btn-outline btn-error btn-xs align-middle"
+                  [disabled]="extraPackage.busy"
+                  (click)="unInstallPackage(extraPackage)"
+                >
+                  @if (extraPackage.busy) {
+                    <span class="loading loading-spinner loading-sm"></span>
+                  }
+                  {{ 'AREAS.INSTALL-SETUP-UPGRADE.COMPONENTS.EXTRA-PACKAGES.ACTIONS.UNINSTALL-NOW' | translate }}
+                </button>
               }
             </h3>
           </div>